ReconSpy-main/requirements.txt
requests
whois
dnspython
colorama
shodan
python-nmap
flask